+++ typings/index.test-d.ts
@@ -1,9 +1,7 @@
 import * as commander from './index';
 import { expectType } from 'tsd';
 
-// We are are not just checking return types here, we are also implicitly checking that the expected syntax is allowed.
-
-/* eslint-disable @typescript-eslint/no-empty-function */
+// We are not just checking return types here, we are also implicitly checking that the expected syntax is allowed.
 
 const program: commander.Command = new commander.Command();
 // @ts-expect-error Check that Command is strongly typed and does not allow arbitrary properties
@@ -148,6 +146,9 @@ expectType<commander.Command>(
 // action
 expectType<commander.Command>(program.action(() => {}));
 expectType<commander.Command>(program.action(async () => {}));
+program.action(function () {
+  expectType<typeof program>(this);
+});
 
 // option
 expectType<commander.Command>(program.option('-a,--alpha'));
@@ -369,6 +370,10 @@ expectType<{ operands: string[]; unknown: string[] }>(
   program.parseOptions(['node', 'script.js', 'hello']),
 );
 
+// save/restore state
+expectType<void>(program.saveStateBeforeParse());
+expectType<void>(program.restoreStateBeforeParse());
+
 // opts
 const opts = program.opts();
 expectType<commander.OptionValues>(opts);
@@ -436,15 +441,12 @@ expectType<commander.Command>(program.executableDir(__dirname));
 expectType<string | null>(program.executableDir());
 
 // outputHelp
-// eslint-disable-next-line @typescript-eslint/no-invalid-void-type, @typescript-eslint/no-confusing-void-expression
 expectType<void>(program.outputHelp());
-// eslint-disable-next-line @typescript-eslint/no-invalid-void-type, @typescript-eslint/no-confusing-void-expression
 expectType<void>(
   program.outputHelp((str: string) => {
     return str;
   }),
 );
-// eslint-disable-next-line @typescript-eslint/no-invalid-void-type, @typescript-eslint/no-confusing-void-expression
 expectType<void>(program.outputHelp({ error: true }));
 
 // help
@@ -542,11 +544,16 @@ expectType<commander.Command>(
     writeErr: (str: string) => {
       console.error(str);
     },
-    getOutHelpWidth: () => 80,
-    getErrHelpWidth: () => 80,
     outputError: (str: string, write: (str: string) => void) => {
       write(str);
     },
+
+    getOutHelpWidth: () => 80,
+    getErrHelpWidth: () => 80,
+
+    getOutHasColors: () => true,
+    getErrHasColors: () => true,
+    stripColor: (str) => str,
   }),
 );
 
@@ -556,7 +563,11 @@ const helperCommand = new commander.Command();
 const helperOption = new commander.Option('-a, --all');
 const helperArgument = new commander.Argument('<file>');
 
+helper.prepareContext({});
+helper.prepareContext({ helpWidth: 120, error: true, outputHasColors: false });
+
 expectType<number | undefined>(helper.helpWidth);
+expectType<number>(helper.minWidthToWrap);
 expectType<boolean>(helper.sortSubcommands);
 expectType<boolean>(helper.sortOptions);
 expectType<boolean>(helper.showGlobalOptions);
@@ -581,7 +592,31 @@ expectType<number>(helper.longestGlobalOptionTermLength(helperCommand, helper));
 expectType<number>(helper.longestArgumentTermLength(helperCommand, helper));
 expectType<number>(helper.padWidth(helperCommand, helper));
 
-expectType<string>(helper.wrap('a b c', 50, 3));
+expectType<number>(helper.displayWidth('some string'));
+expectType<string>(helper.boxWrap('a b c', 50));
+expectType<string>(
+  helper.formatItem('--example', 12, 'example description', helper),
+);
+expectType<boolean>(helper.preformatted('a\nb c'));
+
+expectType<string>(helper.styleTitle('Usage:'));
+
+expectType<string>(helper.styleUsage('foo [options] <file>'));
+expectType<string>(helper.styleCommandText('foo'));
+
+expectType<string>(helper.styleCommandDescription('description'));
+expectType<string>(helper.styleOptionDescription('description'));
+expectType<string>(helper.styleSubcommandDescription('description'));
+expectType<string>(helper.styleArgumentDescription('description'));
+expectType<string>(helper.styleDescriptionText('description'));
+
+expectType<string>(helper.styleOptionTerm('-a, --all'));
+expectType<string>(helper.styleSubcommandTerm('bar [options]'));
+expectType<string>(helper.styleArgumentTerm('<file>'));
+
+expectType<string>(helper.styleOptionText('-a, --all'));
+expectType<string>(helper.styleSubcommandText('bar'));
+expectType<string>(helper.styleArgumentText('<file>'));
 
 expectType<string>(helper.formatHelp(helperCommand, helper));